refactor: types 파일에는 전역 타입만 두기
