--- fe/src/types/audioTypes.ts
@@ -1,5 +1,0 @@
-export interface AudioStreamSetup {
-  audioContext: AudioContext;
-  source: MediaStreamAudioSourceNode;
-  gainNode: GainNode;
-}
