+++ fe/src/types/playerTypes.ts
@@ -4,11 +4,3 @@ export interface PlayerProps {
   isAudioOn: boolean;
   isReady?: boolean;
 }
-
-export interface PlayerListProps {
-  players: PlayerProps[];
-}
-
-export interface AudioControlProps {
-  isOn: boolean;
-}
