+++ fe/src/types/roomTypes.ts
@@ -1,12 +1,3 @@
-export interface RoomDialogProps {
-  open: boolean;
-  onOpenChange: (open: boolean) => void;
-}
-
-export interface JoinDialogProps extends RoomDialogProps {
-  roomId: string;
-}
-
 export interface Room {
   roomId: string;
   roomName: string;
@@ -15,27 +6,7 @@ export interface Room {
   status: 'waiting' | 'playing';
 }
 
-export interface GameRoomProps {
-  room: Room;
-  onJoinRoom: (roomId: string) => void;
-}
-
-export interface PaginationProps {
-  currentPage: number;
-  totalPages: number;
-  onPageChange: (page: number) => void;
-}
-
-export interface JoinGameRoomResult {
-  room: Room;
-  stream: MediaStream;
-}
-
-export interface RoomStore {
-  rooms: Room[];
-  currentRoom: Room | null;
-  refreshRooms: () => Promise<void>;
-  addRoom: (roomName: string, hostNickname: string) => Promise<string>;
-  joinGameRoom: (roomId: string, playerNickname: string) => void;
-  updateCurrentRoom: (room: Room) => void;
+export interface RoomDialogProps {
+  open: boolean;
+  onOpenChange: (open: boolean) => void;
 }
